+++ src/app/core/app-core.module.ts
@@ -47,8 +47,8 @@ const appCoreRoutes: Routes = [
   { path: "app/ex", component: ExComponent }, // TODO: Remove! Only here to test appErrorHandler on production
 
   // Users
-  { path: ":username", component: ProfileComponent, resolve: { title: DynamicTitleResolve } },
   { path: ":username/:timeline", component: ProfileComponent, resolve: { title: DynamicTitleResolve } },
+  { path: ":username", component: ProfileComponent, resolve: { title: DynamicTitleResolve } },
 ];
 
 const coreConfig: ICoreConfig = {