+++ src/app/core/components/history.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { FormControl } from "@angular/forms";
 import { MatDialog } from "@angular/material";
-import { Project, AuthService, ProjectService, Element, ElementItem, ElementField, ElementCell, NotificationService } from "@forcrowd/backbone-client-core";
+import { Project, AuthService, ProjectService, Element, ElementItem, ElementField, ElementCell, NotificationService, ElementFieldDataType } from "@forcrowd/backbone-client-core";
 import { finalize } from "rxjs/operators";
 
 // Component
@@ -66,16 +66,27 @@ export class HistoryComponent implements OnInit {
     }
   }
 
+  get selectedElementLikeCountSet(): ElementCell[] {
+    return this.fields.selectedElementLikeCountSet;
+  }
+  set selectedElementLikeCountSet(value: ElementCell[]) {
+    if (this.fields.selectedElementLikeCountSet !== value) {
+      this.fields.selectedElementLikeCountSet = value;
+    }
+  }
+
   private fields: {
     selectedElement: Element,
     selectedElementField: ElementField,
     selectedElementCellSet: ElementCell[],
     selectedCell: ElementCell,
+    selectedElementLikeCountSet: ElementCell[]
   } = {
       selectedElement: null,
       selectedElementField: null,
       selectedElementCellSet: null,
-      selectedCell: null
+      selectedCell: null,
+      selectedElementLikeCountSet: null
     }
 
   constructor(private authService: AuthService,
@@ -124,6 +135,22 @@ export class HistoryComponent implements OnInit {
           StringValue: this.entry,
         });
 
+        /* Like - Dislike */
+
+        // Item
+        const likeItem = this.projectService.createElementItem({
+          Element: this.selectedElement,
+          Name: `likes ${ this.selectedElement.ElementFieldSet[1].ElementCellSet.length + 1 }`
+        }) as ElementItem;
+        console.log("field ",this.selectedElement.ElementFieldSet[1]);
+        // Cell
+        const likeCell = this.projectService.createElementCell({
+          ElementField: this.selectedElement.ElementFieldSet[1],
+          ElementItem: likeItem
+        });
+
+        this.projectService.createUserElementCell(likeCell, 0);
+
         this.projectService.saveChanges().subscribe(() => {
           this.entry = "";
           this.loadProject(this.project.Id);
@@ -160,8 +187,8 @@ export class HistoryComponent implements OnInit {
     const elementField = this.projectService.createElementField({
       Element: element,
       Name: this.entry,
-      DataType: 1,
-      SortOrder: this.project.ElementSet.length + 1
+      DataType: ElementFieldDataType.String,
+      SortOrder: 1
     }) as ElementField;
 
     // Item
@@ -174,8 +201,34 @@ export class HistoryComponent implements OnInit {
     this.projectService.createElementCell({
       ElementField: elementField,
       ElementItem: elementItem,
-      StringValue: "First"
+      StringValue: "First",
     });
+
+    /* --- */
+
+    // Like Dislikes
+    const elementField2 = this.projectService.createElementField({
+      Element: element,
+      Name: "likes",
+      DataType: ElementFieldDataType.Decimal,
+      UseFixedValue: false,
+      RatingEnabled: false,
+      SortOrder: 0
+    }) as ElementField;
+
+    // Item
+    const elementItem2 = this.projectService.createElementItem({
+      Element: element,
+      Name: "likes"
+    }) as ElementItem;
+
+    // Cell
+    const cell2 = this.projectService.createElementCell({
+      ElementField: elementField2,
+      ElementItem: elementItem2
+    });
+
+    this.projectService.createUserElementCell(cell2, 0);
 
     this.projectService.saveChanges().subscribe(() => {
       this.selectedTab.setValue(this.project.ElementSet.length + 1);
@@ -220,9 +273,11 @@ export class HistoryComponent implements OnInit {
 
       this.projectService.removeElement(element);
       this.projectService.saveChanges().pipe(
-        finalize(()=> {
-          this.selectTimeline(0);
-          this.selectedTab.setValue(0);
+        finalize(() => {
+          if (this.project.ElementSet.length > 0 ) {
+            this.selectTimeline(0);
+            this.selectedTab.setValue(0);
+          }
           this.notificationService.notification.next("Your timeline has been removed!");
         })
       ).subscribe();
@@ -233,7 +288,7 @@ export class HistoryComponent implements OnInit {
   editTimelineHeader(element: Element): void {
     this.notificationService.notification.next("Please write new timeline name form to input then submit");
     this.changeElementName = true;
-    this.entry =this.selectedElement.Name;
+    this.entry = this.selectedElement.Name;
   }
 
   // Edit item
@@ -254,18 +309,32 @@ export class HistoryComponent implements OnInit {
         // Project History
         this.project = project;
 
-        // Selected element
-        this.selectedElement = this.project.ElementSet[this.selectedTimeline];
+        if (this.project.ElementSet.length > 0) {
+          // Selected element
+          this.selectedElement = this.project.ElementSet[this.selectedTimeline];
 
-        // ElementField
-        this.selectedElementField = this.selectedElement.ElementFieldSet[0];
+          // ElementField
+          this.selectedElementField = this.selectedElement.ElementFieldSet[0];
 
-        // ElementCellSet
-        this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
+          // ElementLikeCountSet (ElementCellSet)
+          // for Like - Dislike
+          this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
 
-        this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
+          // ElementCellSet
+          this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
+          this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
+        }
+
         this.isBusy = false;
       });
+  }
+
+  changeLikeCount(value: number, index: number): void {
+    this.isBusy = true;
+    this.selectedElementLikeCountSet[index].UserElementCellSet[0].DecimalValue += value;
+    this.projectService.saveChanges().subscribe(()=> {
+      this.isBusy = false;
+    });
   }
 
   // Set selected timeline element
@@ -274,6 +343,7 @@ export class HistoryComponent implements OnInit {
     this.selectedTimeline = value;
     this.selectedElement = this.project.ElementSet[value];
     this.selectedElementField = this.selectedElement.ElementFieldSet[0];
+    this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet; // for Like - Dislike
     this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
     this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
     this.isBusy = false;